# internal/api/types_gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.3.0 DO NOT EDIT.
package api

// CreateTaskRequest defines model for CreateTaskRequest.
type CreateTaskRequest struct {
	// Desc 更新後の名前
	Desc *string `json:"desc,omitempty"`
	Id   *string `json:"id,omitempty"`
}

// CreateTaskResponse defines model for CreateTaskResponse.
type CreateTaskResponse struct {
	// Desc 更新後の名前
	Desc string `json:"desc"`

	// ID ユーザID
	ID string `json:"id"`
}

// GetDownloadUrlResponse defines model for GetDownloadUrlResponse.
type GetDownloadUrlResponse struct {
	Url *string `json:"url,omitempty"`
}

// UploadExampleRequest defines model for UploadExampleRequest.
type UploadExampleRequest struct {
	EventId     *string `json:"eventId,omitempty"`
	OrgCspDocId *string `json:"orgCspDocId,omitempty"`
}

// UploadExampleResponse defines model for UploadExampleResponse.
type UploadExampleResponse = map[string]interface{}

// GetDownloadUrlParams defines parameters for GetDownloadUrl.
type GetDownloadUrlParams struct {
	// Path ファイルパス
	Path string `form:"path" json:"path"`
}

// CreateTaskJSONRequestBody defines body for CreateTask for application/json ContentType.
type CreateTaskJSONRequestBody = CreateTaskRequest

// UploadExampleJSONRequestBody defines body for UploadExample for application/json ContentType.
type UploadExampleJSONRequestBody = UploadExampleRequest